feat: add signup
This commit adds the signup endpoint that accepts username, name, email
and password.

Adds JWT and MySQL dependencies and sets up a basic connection to the
database.

Adds a HashUtil to generate a SHA256 hash.

Also adds the initial SQL needed in `database.sql`

Also renames the project from `opensource-backend-api` to
`social-media-backend`.