refactor: update profile to user and clean up
This commit updates the naming from "Profile" to "User" in the
controller and service.

This commit also moves the "user exists" checks to user service and
removes the duplicate. Also separates the methods taking username and
email which will be used in the future to separate the conflict
responses.

This commit removes the root controller since it was only used for
testing.

Refs: #20