feat: separate conflict type by body
This commit returns a body when receiving a `409 Conflict` during
signing up. The body contains a `code` that can either be
`auth/email-taken` or `auth/username-taken`, this is used to clarify
what exactly is conflicting and can be used on the Front-end for better
UX.

Also reorganizes exceptions in folders related to their scope.

Also refactors the auth service to be much cleaner using a try-finally
statement.

Also marks "user checks" methods as non-static.

Refs: #20