Separate conflict body code during sign up (#21)
* refactor: update profile to user and clean up

This commit updates the naming from "Profile" to "User" in the
controller and service.

This commit also moves the "user exists" checks to user service and
removes the duplicate. Also separates the methods taking username and
email which will be used in the future to separate the conflict
responses.

This commit removes the root controller since it was only used for
testing.

Refs: #20

* feat: separate conflict type by body

This commit returns a body when receiving a `409 Conflict` during
signing up. The body contains a `code` that can either be
`auth/email-taken` or `auth/username-taken`, this is used to clarify
what exactly is conflicting and can be used on the Front-end for better
UX.

Also reorganizes exceptions in folders related to their scope.

Also refactors the auth service to be much cleaner using a try-finally
statement.

Also marks "user checks" methods as non-static.

Refs: #20