refactor: clean up token validation and log out
After rebasing, this commit updates the log out code to use the
`ValidateToken` method to validate the token, this eliminates the need
to check if the user id exists or not, because if it doesn't, the token
would be invalid because it wouldn't have been issued by us.

This commit also updates the log out to use `try-finally` to close the
connection which is much cleaner and safer because `finally` executes
even if there's an exception thrown.

This commit also moves `DoesUserExistByID` to the user service because
it fits there more than the authentication server, and renames it to
`DoesUserExistById` with a lowercase D in d.

This commit also changes the way that `ValidateToken` works, by
returning the user id directly instead of "credentials" with claims,
this eliminates the use to read the claims manually everytime. This also
updates the `PUT /posts` endpoint to accomodate that change and removes
the `GetUserIdFromToken` method in the authentication service.

Refs: #18