fix(token): handle token expiration correctly.
Previously, when tokens expired, an error was thrown that we did not account for. So in order to unify the token fail exceptions, we made our own `InvalidTokenException` that is thrown whenever an exception was thrown in the token validation code. This way we can catch both exceptions under one.