Token Hotfix (#44)
* fix(token): handle token expiration correctly.

Previously, when tokens expired, an error was thrown that we did not account for. So in order to unify the token fail exceptions, we made our own `InvalidTokenException` that is thrown whenever an exception was thrown in the token validation code. This way we can catch both exceptions under one.

* fix(token): add token validation to follows and unfollows

Previously, if the token is invalid while following or unfollowing a user, a 500 will be thrown. This commit fixes that issue by catching the `InvalidTokenException` and throwing a 401 Unauthorized status code.